Refactor: SRP에 맞게 컴포넌트들 분리 (Header, Detail, GlobalStyle)
* Home 페이지 스타일 수정
