+++ src/components/Header.tsx
@@ -4,6 +4,7 @@ import Link from 'next/link';
 
 import styled from '@emotion/styled';
 import SearchRoundedIcon from '@mui/icons-material/SearchRounded';
+import MenuRoundedIcon from '@mui/icons-material/MenuRounded';
 
 import { common } from '@/styles/common';
 
@@ -15,9 +16,20 @@ export default function Header() {
       </Link>
 
       <nav>
-        <span>영화</span>
-        <span>TV&nbsp;시리즈</span>
-        <SearchRoundedIcon />
+        <div className='drop_movie'>
+          <span>영화</span>
+          <div className='movie_content'>영화</div>
+        </div>
+
+        <div className='drop_tv'>
+          <span>TV&nbsp;시리즈</span>
+          <div className='tv_content'>시리즈</div>
+        </div>
+
+        <div className='drop_search'>
+          <SearchRoundedIcon />
+          <div className='search_content'>검색</div>
+        </div>
       </nav>
     </StyledHeader>
   );
@@ -37,7 +49,9 @@ const StyledHeader = styled.header`
   padding: 0 5rem;
   background: ${common.color.gray01};
 
-  @media screen and (max-width: 390px) {
+  @media screen and (max-width: 600px) {
+    padding: 0 1rem;
+
     a img {
       width: 100px;
     }
@@ -47,14 +61,31 @@ const StyledHeader = styled.header`
     display: flex;
     align-items: center;
 
-    span,
-    svg {
-      padding: 1rem;
-      color: ${common.color.white};
-      cursor: pointer;
+    div.drop_movie,
+    div.drop_tv,
+    div.drop_search {
+      height: 100%;
 
-      :hover {
-        color: ${common.color.gray_hover};
+      span,
+      svg {
+        padding: 1rem;
+        color: ${common.color.white};
+        cursor: pointer;
+
+        :hover {
+          color: ${common.color.gray_hover};
+        }
+      }
+
+      div {
+        display: none;
+        position: fixed;
+        top: 66px;
+        left: 0;
+        right: 0;
+        height: 50%;
+        background: ${common.color.gray01};
+        z-index: 1;
       }
     }
   }
