+++ src/containers/HeaderContainer.tsx
@@ -1,0 +1,5 @@
+import Header from '@/components/Header';
+
+export default function HeaderContainer() {
+  return <Header />;
+}
